[AIRFLOW-5782] Migrate AWS Lambda to /providers/aws [AIP-21]
