+++ docs/operators-and-hooks-ref.rst
@@ -309,7 +309,7 @@ These integrations allow you to perform various operations within the Amazon Web
      - :mod:`airflow.providers.aws.operators.athena`
      - :mod:`airflow.providers.aws.sensors.athena`
 
-   * - `AWS Batch <https://aws.amazon.com/athena/>`__
+   * - `AWS Batch <https://aws.amazon.com/batch/>`__
      -
      - :mod:`airflow.contrib.operators.awsbatch_operator`
      -
@@ -348,8 +348,8 @@ These integrations allow you to perform various operations within the Amazon Web
      -
      -
 
-   * - `AWS Lambda <https://aws.amazon.com/kinesis/>`__
-     - :mod:`airflow.contrib.hooks.aws_lambda_hook`
+   * - `AWS Lambda <https://aws.amazon.com/lambda/>`__
+     - :mod:`airflow.providers.aws.hooks.lambda_function`
      -
      -
 
@@ -387,7 +387,7 @@ These integrations allow you to perform various operations within the Amazon Web
      - :mod:`airflow.contrib.operators.sns_publish_operator`
      -
 
-   * - `Amazon Simple Queue Service (SQS) <https://aws.amazon.com/sns/>`__
+   * - `Amazon Simple Queue Service (SQS) <https://aws.amazon.com/sqs/>`__
      - :mod:`airflow.providers.aws.hooks.sqs`
      - :mod:`airflow.providers.aws.operators.sqs`
      - :mod:`airflow.providers.aws.sensors.sqs`
