+++ scripts/ci/pylint_todo.txt
@@ -266,7 +266,6 @@
 ./airflow/www/widgets.py
 ./scripts/perf/scheduler_ops_metrics.py
 ./tests/contrib/hooks/test_aws_hook.py
-./tests/contrib/hooks/test_aws_lambda_hook.py
 ./tests/contrib/hooks/test_azure_container_instance_hook.py
 ./tests/contrib/hooks/test_azure_cosmos_hook.py
 ./tests/contrib/hooks/test_cassandra_hook.py
